Day 4 Part 1: Rooms and their checksums
That's a fun chain of operations, we take the chars of the room, drop
the dashes, count them up into a hash, sort the hash based on values
with the key as the tiebreak, 5 of those, and collect back into a
string.

We're using beta just now to get rust 1.65, as that's when let else was
stabilised.  It allows us to write:

  let Some(foo) = bar() else { unreachable!(); };
  println!(foo);

Rather than:

  if let Some(foo) = bar() {
       println!(foo);
  }
  else {
       unreachable!();
  }

Which for aoc when we're doing a lot of infallible string parsing is a
bit of a win on clarity.  See Room#From<&str>.from for an example.