Day 11 Part 1: Factory Lift Simulation
It wouldn't be advent without a graph traversal game.  This one has no
explicit cost function, the number of moves is the thing to minimise so
use bfs from the pathfinding crate.

Also bring in the regex crate, as parsing the input by hand was getting
tedious, and lazy_static to memoize the regex compilations.