Day 19 Part 1: Josephus is a very naughty elf
I owe all this to the Numberphile video on the Josephus problem.

I'm indulging in a little light microbenchmarking to see if the version
that slices and makes new strings is faster/slower that the version that
rotates the bytes in place.  It's a wash when running from the CLI:

  AOC 2016
  Day 19 - Part 1 - rotate : 1815603
          generator: 5.493µs,
          runner: 12.378µs

  Day 19 - Part 1 - slice : 1815603
          generator: 407ns,
          runner: 11.36µs

Criterion hands it to the rotate version, with ~260ns vs ~550ns, which
makes sense as it avoids an allocation.  Both seem acceptably fast for
our purposes though.