Day 19 Part1: Correctly optimised
Rather than optimising things that make strings of bit representaions,
we can actually mess around in the bits by calling on leading_zeros()
and doing the bitmasking to hide the top bit.

This is 4.6 nanoseconds per cycle on my machine, a clear win over doing
the stupid.  It was a little harder to write though as I had to find the
documentation for usize and grok that leading_zeros would get me the
leading bit.