Day 19 Part 2: More maths
Inspired by the numberphile video on how to do the analysis, we did one.
Took a while to spot the pattern.

 P  | W
 ------
 1  | 1
 2  | 1
 3  | 2
 4  | 1
 5  | 2
 6  | 3
 7  | 4
 8  | 5
 9  | 6
10  | 1

It's advancing in ones, and resets just after a power of 3.

This guided guessing to get us to W = P - 3^m, which fits the general
pattern of the Josephus formula, and the aoc site confirms our answer.