Day 22 Part 1: Add tgl instruction to assembunny
We now make the program memory be owned by the CPU so that tgl can
mutate the program state.

Also we soften Cpy to be Cpy(Value,Value) as with Jnz() so they can be
swapped.  We can only Cpy to Registers, so have the evaluator match on
Cpy(Value, Register) as distinct from Cpy(Value, Literal)