Refactor the flow of day24::solver
Remembering the technique from day 18 of using concat, we can make the
second data into a fairly clear iterator chain, which can do the min()
for us.

I think I'd prefer to be able to do this with .chain() but things along
the lines of building the iterator up dynamically don't work as each
iterator function returns its own Struct type.

eg:
    std::iter::once(start)
       .chain(middle)
       .chain(if go_back { std::iter::once() } else { std::iter::none() })

Can't work as the two arms of the if statement return Once and None
structs.

Hmm, maybe if we threw some `dyn impl Iterator` casts in there it would
work better.